Recover handler renaming
